Admin UI: Extract shared withRouter decorator using @wordpress/route private APIs
Replace direct @tanstack/react-router imports with a shared withRouter
decorator that uses @wordpress/route private APIs, keeping consistent
with Gutenberg conventions.

See: https://github.com/WordPress/gutenberg/pull/76467#issuecomment-4060951002
